+++ src/Fable.Cli/Pipeline.fs
@@ -82,15 +82,7 @@ module Js =
     type BabelWriter
         (com: Compiler, cliArgs: CliArgs, pathResolver: PathResolver, sourcePath: string, targetPath: string)
         =
-        // In imports *.ts extensions have to be converted to *.js extensions instead
-        let fileExt =
-            let fileExt = cliArgs.CompilerOptions.FileExtension
-
-            if fileExt.EndsWith(".ts", StringComparison.Ordinal) then
-                Path.ChangeExtension(fileExt, ".js")
-            else
-                fileExt
-
+        let fileExt = cliArgs.CompilerOptions.FileExtension
         let sourceDir = Path.GetDirectoryName(sourcePath)
         let targetDir = Path.GetDirectoryName(targetPath)
         let memoryStream = new IO.MemoryStream()