+++ src/fable-library-ts/Int32.ts
@@ -1,5 +1,5 @@
-import { FSharpRef } from "./Types.js";
-import { Exception } from "./Util.js";
+import { FSharpRef } from "./Types.ts";
+import { Exception } from "./Util.ts";
 
 export type int8 = number;
 export type uint8 = number;